requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.29.2
azure-identity==1.14.0
azure-monitor-ingestion==1.0.2
Babel==2.12.1
blinker==1.6.2
Bootstrap-Flask==2.2.0
boto3==1.28.29
botocore==1.31.29
cachelib==0.9.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
checkdmarc==4.7.0
click==8.1.5
colorama==0.4.6
crontab==1.0.1
cryptography==41.0.3
dateparser==1.1.8
dnspython==2.3.0
dominate==2.8.0
elasticsearch==7.13.4
elasticsearch-dsl==7.4.0
email-validator==2.0.0.post2
expiringdict==1.2.2
Flask==2.3.2
flask-babel==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.0.2
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
flask-mongoengine==1.0.0
Flask-Session==0.5.0
flask-sock==0.6.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
freezegun==1.2.2
frozenlist==1.4.0
geoip2==4.7.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
html2text==2020.1.16
httplib2==0.22.0
humanize==4.7.0
idna==3.4
IMAPClient==2.3.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kafka-python==2.0.2
lxml==4.9.3
mail-parser==3.15.0
mailsuite==1.9.15
Mako==1.2.4
MarkupSafe==2.1.3
maxminddb==2.4.0
mongoengine==0.27.0
msal==1.23.0
msal-extensions==1.0.0
msgraph-core==0.2.2
multidict==6.0.4
oauthlib==3.2.2
packaging==23.1
parsedmarc==8.6.1
portalocker==2.7.0
protobuf==4.24.0
publicsuffix2==2.20191221
publicsuffixlist==0.10.0.20230814
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyleri==1.4.2
pymongo==4.4.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rq==1.15.1
rq-scheduler==0.13.1
rsa==4.9
s3transfer==0.6.2
simple-websocket==0.10.1
simplejson==3.19.1
six==1.16.0
SQLAlchemy==2.0.18
timeout-decorator==0.5.0
tqdm==4.66.1
typing_extensions==4.7.1
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
visitor==0.1.3
Werkzeug==2.3.6
wsproto==1.2.0
WTForms==3.0.1
xmltodict==0.13.0
yarl==1.9.2